+++ package.json
@@ -43,6 +43,7 @@
     "highlight.js": "^9.15.6",
     "lint-staged": "^8.1.5",
     "markdown-it": "^8.4.2",
+    "markdown-it-json": "^1.0.2",
     "markdown-it-link-attributes": "^2.1.0",
     "markdown-it-mark": "^2.0.0",
     "markdown-it-regexp": "^0.4.0",
